requirements.txt
Flask
pandas
pymysql
lxml
cryptography
datetime
bs4
urllib3
mysql-connector-python
